pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "argparse>=1.4.0",
    "numpy>=2.2.4",
    "paho-mqtt>=2.1.0",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "streamlit>=1.44.1",
]
